The new routines that I'm adding constitute a package making it possible to use the RANGS/GSHHS database from NCAR Graphics.  A few changes were required in the BLOCK DATA and in the parameter-access routines.
There are some known problems with these routines, but they mostly
have to do with drawing small-scale maps (those that show all or most
of the globe) or with drawing maps on which a limb line is visible,
neither of which is a likely situation, given the highly detailed
nature of the data.

Documentation of these routines will be forthcoming next week.

It remains to be decided where the datasets will reside, since
they are so big as to approximately triple the size of an NCAR
Graphics installation.  There is a routine that I call to retrieve
the directory where the data sit; for the moment, the default
version of that routine logs an error and tells the user that he
or she must supply his or her own version of the routine ...

A work in progress ...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6254 10fe2890-8a40-4696-8b1d-7ccbb19c398e