This is being done in response to a problem reported by Livio Casella.
I have modified these routines to insert a call to SFLUSH prior to each
call to SET, so that anything in PLOTIT's line-draw buffers will be sent
to the metafile prior to changing the clipping rectangle.

Actually, this is an unnecessary change except on a machine like Livio's,
which, for some reason, rounds in such a way as to effectively move an
axis line outside the grid window.  Even then, that wouldn't be a problem
except that, if that axis line is still sitting in PLOTIT's buffers when
the final SET call is done for the user, it ends up getting clipped ...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6265 10fe2890-8a40-4696-8b1d-7ccbb19c398e