I am adding an example that shows how to use the new EZMAP routines that access the RANGS/GSHHS database of very-high-resolution coastlines.  This example also shows how to use routines that will draw lat/lon labels along an arbitrary line on a map.
Note that, as it stands, this example will fail the first time it tries to
access the new database.  This is because the user needs to download the
RANGS/GSHHS data and supply a subroutine MDRGDI telling EZMAP where the
data have been placed.  Instructions are given in the code and I think I
need to put some prints in the program telling the user what to do; back
in a minute ...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6311 10fe2890-8a40-4696-8b1d-7ccbb19c398e