Hmmmmmm.  I didn't actually change every file in this directory.  Most of the changes were in the routines that deal with the RANGS/GSHHS database, with a few changes in the parameter-access routines and the BLOCK DATA. I guess the others appear to have been changed only because I changed the line with the $Id$ on it ...
In any case, the principal object of these changes is to incorporate all
the new code that allows for fill using AREAS, which allows for arbitrary
transparency of each of the five classes of filled objects.  Secondarily,
I have implemented a routine that edits certain polygons in the RANGS/GSHHS
database to remove self intersections that otherwise give AREAS fits.  (In
two places (one in Russia and the other off the coast of Chile), there were
errors of other types.

I have also added some new arguments to the routines MDRGOL and MDRGSF; all
scratch space required must now be supplied by the user.  This is not only
to keep about 3 megabytes of memory from sitting idle, but to make possible
some future uses of the data that will require user access to the scratch
space.

Just realized that I forgot to "add" six new routines ... sigh ...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6386 10fe2890-8a40-4696-8b1d-7ccbb19c398e