1) Added a new resource to MapPlot: mpDataResolution. This only applies when the mpDataBaseVersion is set to RANGS_GSHHS (or one of its aliases: RANGS or HighRes). It allows the users to explicitly set one of the 5 available resolution levels for this database. If set to the default value, "Unspecified" or equivalent alias "UnspecifiedResolution", the resolution level will be determined dynamically.
2) Modified the algorithm for culling overlapping labels such that it
now sets a uniform stride in both directions from a dynamically determined
pivot label string chosen to be among the "roundest" values in the label set.
If the end labeling is turned on (LabelBar only) then the end labels
are always shown regardless of the stride. Because of this change in
behavior the name of the resource was changed from CullLabelOverlaps to
LabelAutoStride.

3) Fixed a bug in MapV41DataHandler.c that disallowed setting the
Ocean fill color when the dataset is "Earth..2".

4) Fixed ContourPlot, VectorPlot, and StreamlinePlot bugs involving the
LabelBar end labels. If the levels were set such that their range exceeds
the range between the min and max data levels, the end labels were still
incorrectly showing the min and max data levels.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6417 10fe2890-8a40-4696-8b1d-7ccbb19c398e