While reviewing the code of this example for possible future use in code to deal with "curvilinear grids", I found that it included a version of the function ISINPO (which tells you whether a given point is part of the interior of a given polygon or not) which could give incorrect results in certain cases.  (None of those cases actually arise in the example, but they could easily arise in an adaptation of it).  To correct this, I have swapped in a more robust version of ISINPO.
Unfortunately, this version is a little slower that the other one.  I have
yet another version which is of intermediate speed, but it requires that the
points of the polygon be defined in counterclockwise order.  Again, that's
easy to ensure in the context of this example, but would not necessarily be
easy to ensure in a modified version of the example implementing a different
curvilinear grid.

I'm still looking into this, so I may come up with yet another version of
ISINPO which is both as fast as the original and demonstrably correct.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6426 10fe2890-8a40-4696-8b1d-7ccbb19c398e