I have added some new code to CPHLLB that enables it, on user request, to search for highs and lows that were not found by the original code - those at grid points where the field value is exactly the same as that at one or more of the adjacent grid points.  Normally, the probability of this would seem to be low, but, if the field has been subjected to a packing/unpacking operation that discretizes the data, the probability of a problem is increased.
Because it can take some time to do the new search for highs and lows,
it is not done automatically, but only if evidence is found, during the
"normal" search, that it might be justified and then only if the user
has set a new internal parameter 'HLE' non-zero.  Documentation of all
this will be forthcoming.

The problem this fixes was reported by Greg Thompson via Dave Brown.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6435 10fe2890-8a40-4696-8b1d-7ccbb19c398e