This is a first-cut implementation of curvilinear coordinates. The only plot object with support so far is ContourPlot. Graphics primitives draw using the Curvilinear Transformation may work to a certain extent, but only if no clipping is involved.
ContourPlot lines, fill and raster fill are supported, but note that
only point data is supported, not data specified using cell boundaries.
Also the 2D coordinate arrays must be samples of a basically continuous
coordinate field. Modular coordinates must be preprocessed to remove
discontinuities resulting from locations being specified on both sides
of a modular boundary.

To use the capability, simply set the ScalarField resources sfXArray
and sfYArray using 2D arrays instead of 1D arrays. The 2D arrays must
have the same dimension sizes as the data array provided to the
sfDataArray resource. There are no sub-setting capabilities for
the 2D coordinate arrays so do not use the sf{X|Y}C{Start|End}{Index|SubsetV}
resources. The sf{X|Y}C{Start}{End}V resources are mapped into read-only
resources that specify the max and min values of the X and Y coordinate
arrays.

Once the 2D coordinate arrays have been supplied, behavior is analagous to
the rectangular Irregular coordinate facility, except that a curvilinear
plot that is not overlaid does not support TickMarks currently. Practically
speaking, you will most often want to overlay the curvilinear data on
a MapPlot or possibly on a LogLinPlot. In both of these cases the
curved boundaries of the data space will become apparent, except
in the case of a MapPlot with the projection resources set to match the
native projection and limits of the curvilinear grid.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6506 10fe2890-8a40-4696-8b1d-7ccbb19c398e