Removed all references to the function "SECOND", which was intended for timing purposes, has probably never been used by anyone, and was causing problems for Mary on certain systems.
The routine "CAPSAP", in the file "cpexcc.f", no longer has an argument
"TIME" and all calls to it have been changed to remove this argument.
It is possible that this may cause trouble for some users; I have seen
a few user codes, adapted from the examples, in which this routine was
called.  In all cases, I think, those user codes also contained a local
version of the routine CAPSAP, so they would not be affected, but I can
imagine this change causing some confusion for some users ...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6534 10fe2890-8a40-4696-8b1d-7ccbb19c398e