Added a new 'SphericalTransformation' to implement Dave Kennison's new 2D coordinate method that works for lat-lon grids such as the POP grid.
SphericalGeometry.c is a adaptation of the C version of
the support routines that accompany cpex16.

SphericalTransObj.c implements the new Spherical Transformation.
The polyline and polygon routines have not been worked on. Attempts
to draw graphics primitives through this transformation are likely
to blow up.

There is a new enumerative type at the DataItem level:
 typedef enum _NhlGridType {
       NhlBASICGRID,
       NhlSPHERICALGRID
 } NhlGridType;

 #define NhlTGridType "GridType"

and a new resource NhlNsfGridType of type NhlGridType has been added
to the ScalarField.

This allows the user to define the 2D grid as a spherical grid -- which
will cause ContourPlot to use the SphericalTransformation, or as a
'basic' grid, which results in the use of the previously implemented
CurvilinearTransformation.

"SphericalGrid" is the default.

ScalarField was modified to permit subsetting 2D coordinate grids by index
only. Stride and coordinate subscripting are not supported. Also the
max and min coordinate values are not particularly meaningful in the
2D coordinate grid context, particularly when there is wrap-around along
the modular longitudinal axis.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6640 10fe2890-8a40-4696-8b1d-7ccbb19c398e