This implements cell-boundary coordinate arrays for ScalarField, cell-boundary based plotting for ContourPlot. Fixes for a problem that was causing unnecessary redraws (when segments could be used) for all the plot objects are also included. Also a fix (kludgy at this point) for a problem involving data min/max X extents for the SphericalTransObj. This last problem really involves the modular property of the longitudinal axis, and is still awaiting a more definitive treatment from the ScalarField on down.
The ScalarField now allows in the 1D coordinate case sfXArray and/or sfYArray
values containing 1 more element than the corresponding sfDataArray dimension.
Given such a coordinate array, ScalarField sets an "is_bound" flag in the
internal data object passed to ContourPlot. ContourPlot then treats the
coordinate as a cell bounding array. This causes the discrete raster
plotting to use the coordinate values as divisions between data cells, rather
than to treat each coordinate value as a cell center.

In the 2D case, both the X and Y dimensions of the coordinate arrays
must have 1 more element than the sfDataArray.

Normal line and fill contouring as well as smoothed raster contouring seem
to work with boundary specified coordinates, although I have not convinced
myself that this will be so in all situations. But for now there is
no restriction on using cell boundary coordinates for any ContourPlot
purpose.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6649 10fe2890-8a40-4696-8b1d-7ccbb19c398e