In the CGM_open call, when the file is opened with a request in "a" mode through fdopen call, it causes a problem on LINUX system and so it is changed to call the fdopen with "w" mode.
The problem is that, on LINUX system, using fdopen with "a" force caller to
write at the end-of-the-file regardless of the seek made to the file. But
on other systems like solaris, SGI, and AIX, it allows to seek on the file
pointer even after opening the file stream using fdopen with "a" mode and
that is why the code was behaving differently on LINUX system causing it to
break.

for example: The old code in CGM_open call in file cgm_tools.c is like
        if (! (fp = fdopen(fd, "a"))) return(-1);

and now its changed to :
        if (! (fp = fdopen(fd, "w"))) return(-1);

so now when one use CGM_open with "a" mode, the file pointer will not point
to the end-of-the-file, instead it will point to the begining of the file
and you have to seek to the end-of-the-file to append at the end of the file.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6672 10fe2890-8a40-4696-8b1d-7ccbb19c398e