Fix for missing accumulation variables reported by Greg Thompson. NCL was not distinguishing between accumulations with different time periods. It saw these records as identical and removed the last ones found as duplicates.
Now if the time range indicator field indicates an accumulation, an
average, or a difference, the time period (difference between fields P2
and P1) is calculated. If the period is 0 (most often the case in
files I have tested) the variable name is left alone, but if the
value is non-zero, the time period value plus a unit indicator is
appended to the name. For the example file that Greg sent, the
accumulation variables now look as follows:

      float NCPCP_236_SFC_acc1h ( gridx_236, gridy_236 )
      float NCPCP_236_SFC_acc3h ( gridx_236, gridy_236 )
      float ACPCP_236_SFC_acc1h ( gridx_236, gridy_236 )
      float ACPCP_236_SFC_acc3h ( gridx_236, gridy_236 )
      float WEASD_236_SFC_acc3h ( gridx_236, gridy_236 )

For files with only a single accumulation, average, or difference time
period, this results in a degree of incompatibility with previous
versions, but I think it is necessary in order to correctly
represent the complete contents of GRIB files in general.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6696 10fe2890-8a40-4696-8b1d-7ccbb19c398e