Fixes for ContourPlot labeling:
cnLabelMasking now works with all values of cnLineLabelPlacementMode.
Label masking behavior has been changed slightly. Now it only
masks contour lines, not contour fill. This seems to be the most
useful behavior. It is no longer necessary to set cnFillDrawOrder
to an earlier draw phase than the line draw in order for masking
to work just on contour lines. To see the effect of cnLabelMasking
it is necessing to change the relevant label background to 'transparent'.

I also fixed some problems involving the transparent 'color' for the
various label background and perimeter resources. There are some
complicated dependencies between the High and Low label resources
involving transparency that are better understood and handled now I think.
(At least GKS error states are avoided.) It needs more formal documentation,
but here is a brief explanation:

Conpack ties a number of High/Low label resources together. For that
reason some resources are controlled only by the cnHighLabel... resource,
and the corresponding cnLowLabel resources are ignored. These include:
  cnLowLabelFontHeightF
  cnLowLabelAngleF
  cnLowLabelPerimSpacingF
  cnLowLabelPerimOn

The cnLowLabelBackgroundColor and cnLowLabelPerimColor *can* be set
independently of the corresponding high label resource as long as
neither of them are set to 'transparent'. However, because the transparent
'color' is simulated by turning off the background or the perimeter at
the Conpack level, and because the on/off switch applies to both, if the
high label resource is set to 'transparent' the low label resource value
is ignored and it also becomes 'transparent'. Conversely, if the high label
resource is set to a non-transparent value the low label resource cannot
be transparent. If it is set 'transparent' it will be defaulted to
'background' in the case of cnLowLabelBackgroundColor and 'foreground' in
the case of cnLowLabelPerimColor. For now this all takes place without
any warning messages. I should probably correct this.

Still to come is a resource for changing label 'density'.
 -dave


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6701 10fe2890-8a40-4696-8b1d-7ccbb19c398e