Fixed a work array that gets passed to the LAPACK routine DGESVD to be:
  max(3*nsvmx+max(ncx,ncy) , 5*nsvmx)

instead of

  max(3*nsvmx+max(ncx,ncy) , 5*nsvmx-4)

There was some confusion about which one it should be, because
the LAPACK version does *not* have the "-4", and the complib.sgimath
library on the SGI does.

Since it's okay for the work array to be larger than the size
required, we went with the larger size to be safe.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6705 10fe2890-8a40-4696-8b1d-7ccbb19c398e