Replaced:
   systemfunc("tmp$$")

type calls with:

  unique_string("tmp")

type calls.  This is to fix a potential problem with using
"systemfunc" which forks a process every time.

"unique_string" is a simple function that implements a static integer
counter, so that you can be assured of getting a unique number every
time (within an NCL session).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6708 10fe2890-8a40-4696-8b1d-7ccbb19c398e