Fixes some remaining problems with the accumulation type variables: Negative accumulations are made positive by modular arithmetic where it makes sense, and simply by taking an absolute value where it doesn't. I have not found any official documentation about this, so I'm guessing. For the more rarely used units I'm probably wrong. But for hours, which is all I have seen for accumulation times so far, I have reasonable confidence it works properly.
Also I put in a bit of code that causes NCL to ignore octet 17 of the
GDS when comparing the GDS of different variables to see if a new
coordinate dimension is required. NCL takes no notice of differences
in this octet, other than to create duplicate coordinates,
only one of which can be accessed. Until NCL is able to properly take
account of the flags stored in this octet, it seems better to ignore
it completely.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6729 10fe2890-8a40-4696-8b1d-7ccbb19c398e