Fix for TT #036248:
This was not a curly vector bug. Actually curly vectors were behaving
properly. Instead the bug was that the ref length for normal line and fill
vectors was not being calculated properly if the ref magnitude was set,
but the ref length resource had it's default value of 0.0 (meaning that
VectorPlot should determine the ref length itself).

Setting the ref magnitude when the ref length is determined by default
should not change the length of any arrows, but simply set the reference
magnitude which is displayed in the vector annotation. The bug was that
the maximum length was improperly affected by changes to the ref
magnitude. The curly vector code was not affected by the bug: hence
the inconsistency that led to the bug report.

Further explanation:
The algorithm for setting the default reference length is simple-minded.
It works okay for uniform grids when no clipping of the data space is
involved. But for grids over map projections, especially when the
viewport shows only a portion of the data, it often results in the
vectors appearing too small. The user in this case must set the
ref length to a larger value.

Once the ref length has a fixed value set by the user, setting the
ref magnitude does legimately affect the maximum size of the vectors.
The maximum size will increase as the ref magnitude is set to smaller
and smaller values.

The documentation should probably recommend that for best results
the reference magnitude should be set to values nearer the maximum
magnitude than the minimum, and that size adjustments be done
primarily using the ref length resource.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6745 10fe2890-8a40-4696-8b1d-7ccbb19c398e