"Fixed" some of the gsn_csm_vector* scripts so that if the user moves the reference annotation up *and* a labelbar is being drawn, then the labelbar doesn't get moved up into the tick mark labels.
This was happening because the labelbar was in a zone one greater than
the ref anno, and when the ref anno was moved, the labelbar tried to
stay with it.  I fixed this by putting the labelbar in the same zone
as the ref anno if the ref anno is moved.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6749 10fe2890-8a40-4696-8b1d-7ccbb19c398e