Addresses TT#037690 Performance enhancements for array variable initialization and assignment.
OpsFuncs.c: the new operator has been enhanced to double the number
of elements initialized (to the fill value) by memcpy for each iteration.
It does this by copying the increasing already initialized memory area
to the remaining uninitialized area.

NclMultiDValData.c: where possible contiguous data is now copied used
single invocations of memcpy. The routines that affect various kinds
of array assignment have been modified including:
 1) unsubscripted assignment where the types differ
 2) unsubscripted = subscripted: t = u(1,:,:)
 3) subscripted = unsubscripted: t(1,:,:) = u
 4) subscripted = subscripted: t(1,:,:) = u(0,0:9,:)
 5) subscripted = scalar: t(1,:,:) = 10.0

The optimization is only possible when certain conditions are met:

The stride is positive; the dimension is not reversed; the
dimension is not reordered, and the dimension does not employ vector
subsetting. This is on a fast to slow basis (right to left in NCL).
Proceeding leftword, the first dimension that does not meet these
critera marks the upper bound of contiguous data. A dimension
that is a proper subset can be contiguous, but dimensions to
left are excluded.

Operation should be transparent to the user except for speed of
execution. The speed difference is only apparent for large
variables.

Some fairly sensitive code has been altered to implement this
enhancement, so beta testing is essential.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6779 10fe2890-8a40-4696-8b1d-7ccbb19c398e