Added ability for scalar variables to be added to NetCDF and HDF files. This is done via the filevardef routine, by setting the dimension name argument to the pre-defined NCL reserved value "ncl_scalar". Note that one should not attempt to define the dimension "ncl_scalar" using filedimdef.
Assuming 'f' references a writable file, here is an example:
    filevardef(f,(/"a","b"/),(/"float","integer"/),"ncl_scalar")

Within NCL, file variables defined this way will have the unitary dimension
"ncl_scalar", but when read by tools such as ncdump, they will be real
scalars. This is the same behavior that NCL has always had for files that
contain pre-defined scalar variables, so it is a natural extension.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6794 10fe2890-8a40-4696-8b1d-7ccbb19c398e