Adds built-in function:                 function isnameddim(                         var,                         dim_nums : integer                 ) var Must be a variable parameter of any type and dimensionality
dim_nums
a scalar or a 1-D array of dimension numbers

Returns True for every element of 'dim_nums' that is the number
of a dimension that is named in variable 'var'. Returns False if
the dimension number has no name or is not a valid dimension number
for the variable. The input values may be in any order and the output
values correspond one for one with the input order. As a special case,
if 'dim_nums' is set to -1, returns a value for each dimension of
'var' , in the order of its dimensions.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6804 10fe2890-8a40-4696-8b1d-7ccbb19c398e