Updates the centers, sub-centers, and models tables using the currently documented values at http://www.nco.ncep.noaa.gov/pmb/docs/on388/ with these exceptions:
Model index 84 is changed from "MESO ETA Model (currently 12 km)"
to simply "MESO ETA Model", because I believe the same index has been
used historically for a number of different model resolutions.
Model indexes 83 and 85 are changed from "No longer used" to
what they were previously: "ETA Model - 80 km version" and
"ETA Model - 30 km version", on the assumption that any GRIB file with
these model indexes would be an old one.

The sub-centers table keeps indexes 150 - 170, which Ethan had
documented (no idea where he got the information), but which are not
listed in the current documentation. Again it seems that if a GRIB file
has these, then it knows about them.

Also created a new struct type GribTable which allows the index and
identifying string to be stored next to each other. This allows for
easier auto-generation scripts, and is much easier to read and keep track
of than Ethan's method of using separate arrays for the indexes and the
strings.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6914 10fe2890-8a40-4696-8b1d-7ccbb19c398e