Added ".grib" as a recognized extension for GRIB files. I may still add ".GRIB" and ".GRB" later.
Fix for TT#039626. This concerns a problem with a script using "fbinrecwrite"
repeatedly in a loop.

The problem was *not* with "fbinrecwrite", but with "systemfunc", which was
also called within the loop.

"systemfunc" opens 2 file descriptors using the "pipe" call. It was
failing to close the read end of the pipe after it finishes reading the
results of the command that is executed. On Solaris and IRIS systems, this caused
the NCL process to run out of file descriptors before the loop completed.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6924 10fe2890-8a40-4696-8b1d-7ccbb19c398e