Adds four new HLU API functions:
int NhlNewDashPattern
(
	int	wid,
	char	*dash_string
)
NhlErrorTypes NhlSetDashPattern
(
	int wid,
	int	index,
	char	*dash_string
)
int NhlNewMarker
(
	int	wid,
	char	*mark_string,
	float	x_off,
	float	y_off,
	float	aspect_adj,
	float	size_adj,
	float   angle
)
NhlErrorTypes NhlSetMarker
(
	int instance,
	int	index,
	char	*mark_string,
	float	x_off,
	float	y_off,
	float	aspect_adj,
	float	size_adj,
	float  angle
)

ContourPlot and MapPlot make copies of the dash table. Since the
table is no longer static it is now updated before each draw.
Now that dash patterns are visible to users, it seemed advisable to
use the characters that dashpack uses ($ and _) rather than the
old characaters used by dashchar ($ and '), because "'" is more
likely in my opinion to be needed in a line label. But Ezmap still
uses Dashchar, so the dash strings are now modified before sending
to Ezmap. Like color tables, the dash and marker tables are now
specific to each workstation.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6985 10fe2890-8a40-4696-8b1d-7ccbb19c398e