Adds new "..FillDotSizeF"  resources for controlling the size of dots in stipple fill patterns. Currently it only applies when the fill pattern is set to 17.
The class resource "FillDotSizeF" can be used in a resource file to
set all instances of this type of resource.

There are four new user-visible instance resources in this category:

gsFillDotSizeF
lbFillDotSizeF
mpFillDotSizeF
cnFillDotSizeF

There are no array type resources, so the fill dot size is always uniform
within a single plot object. As the dot size increases the user must
increase the fill scale factor to avoid overlap. The units are NDC. The
default value of 0.0 causes the dots to be drawn as before, using a
workstation dependent minimum dot size. A caveat is that individual
dots are not clipped around the edges of fill areas -- this becomes
more noticeable as the dot size increases.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6991 10fe2890-8a40-4696-8b1d-7ccbb19c398e