Function    fspan(start [float], finish [float], spacing [int]) changed to:             fspan(start [numeric], finish [numeric], spacing [int])
to accomodate promotion of arguments to type double.  When a user has data of
type double, precision is lost if it is demoted to type float.  If either of
the arguments "start" or "finish" are of type double, computation is done with
type double.  Any other numeric type is promoted to type float for computation.

This change also fixes the problem of integer variables being given as the
"start" and "finish" arguments generating a somewhat cryptic coercion message.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6993 10fe2890-8a40-4696-8b1d-7ccbb19c398e