Modifies the interfaces to NhlNewMarker and NhlSetMarker: the font is now a separate parameter so that function codes are not needed in the the marker string, which is now typically just a single character.
Also changes the behavior of some of the parameters and the way that
markers are produced. Markers are now automatically centered relative
to their bounding box using the 'CE' parameter to plotchar. Therefore
the x_off and y_off function parameters can usually be set to 0.0.
Non-zero values are only needed when a non-centered marker is required.
The units of x_off and y_off are now 'fraction of marker size', which
is in NDC units, if the marker has been 'standardized' with the
proper size adjustment.

The aspect adjustment parameter is now multiplied by the default HLU
aspect ratio of 1.3125. This means that the aspect adjustment parameter
can more often be left at its default value of 1.0.

In order to create a marker that acts like one of the standard markers
the size adjustment parameter needs to be tweaked such that the width of
the marker is equal in NDC units to the set value of the xxMarkerSizeF
resource. For a given value of the xxMarkerSizeF resource and the size
adjustment parameter, the width of the marker remains constant. The
height varies as the aspect adjustment parameter is varied.

The parameters associated with the built-in markers were necessarily
modified because the behavior of the parameters changed. But in the
process, the built-in markers are now more accurately sized and centered
and their aspect ratios are now all square. This is considered a bug-fix,
but it may change the appearance of plots generated by existing code
just a bit. Hopefully this won't cause a big ruckus.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6995 10fe2890-8a40-4696-8b1d-7ccbb19c398e