Fix for TT #040731
In addition to the problem with the time dimension coordinate values, it
turns out that the time dimension was not properly attached to the variables
for the type of CCM file represented by the test case file DFI668.ccm.
I have fixed this and NCL now properly handles the time dimension in this
type of file correctly.

Dennis notes that the NetCDF version of the file produced by ccmtonc seems
to be "off" when read by NCL. I think by this he means that the temperature
contours do not align with the continents in the expected way, producing
"ghost" images of the continents that are offset from the actual continental
outlines. Now that the CCM file is handled correctly it appears to me that
the native CCM file is off in the same way, suggesting that the CCM file
itself contains the error. If NCL is wrong, then it is wrong for both
formats in the same way, which seems somewhat unlikely.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7003 10fe2890-8a40-4696-8b1d-7ccbb19c398e