This is another approach to solving the problem of TT#040246 that concerns assignment with singleton dimensions. Because of the conflict with the dimension and attribute removing property of the  (/ ... /) operation, there is no way to avoid collapsing to a single dimension when assigning to a previously undefined variable, i.e:  x = (/ (/6/), (/2/) /) must still produce a 1D array of size 2, rather than a 2D array of size (1,2).
The approach I have now taken is to relax the assignment operation such
that singleton dimensions are ignored. This allows for literal and/or
variable assignment to variables with singleton dimensions. Basically,
singleton dimensions are ignored during assignment. Therefore

    i = new( (/2,1/), integer)
    i = (/ (/6/), (/2/) /)

works, because the singleton dimension of i is ignored. So actually this
means that
    i = (/ 6, 2 /)
also works, as does
    i = (/(/ 6, 2 /)/)

Dimensions and coordinates are still propagated correctly, even if
ignored singleton dimensions are interleaved. Example:

 t = new((/2,1,2/),float)
 u = new((/2,2/),float)
 u!0 = "zero"
 u!1 = "one"
 u = 4

 t = u

 print(t)

Variable: t
Type: float
Total Size: 16 bytes
            4 values
Number of Dimensions: 3
Dimensions and sizes:	[zero | 2] x [1] x [one | 2]
Coordinates:
Number Of Attributes: 1
  _FillValue :	-999
(0,0,0)	 4
(0,0,1)	 4
(1,0,0)	 4
(1,0,1)	 4


Note that the original problem (other that assignment to undefined variables)
could have been handled using left-hand side subscripting, so this
modification is not really necessary. If there's any problem with it,
I will eliminate it, and we can go back to the original code.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7047 10fe2890-8a40-4696-8b1d-7ccbb19c398e