This check-in addresses TT# 033620 -- improvements in ContourPlot line labelling. I have added a new resource cnLineLabelDensityF, which allows a simple, though imexact, method for controlling the number of line labels on a ContourPlot for both the Randomized and Computed line label placement modes. It has no effect for the Constant mode, which depends on cnLineDashSegLenF to determine the spacing.
If cnLineLabelDensityF is set to its default value, 0.0, it has no effect
and the line label spacing can be controlled using cnConpackParams, as
always.
If cnLineLabelDensityF is set to 1.0, then the default settings of the
appropriate Conpack parameters apply, overriding any that are set by
cnConpackParams.
As cnLineLabelDensityF is set to values greater than 1.0, the number of
labels increases, while setting it to positive values less than 1.0 causes
the number of labels to decrease.

Since I am using a single parameter as the basis for setting a group of
parameters (with complex interactions, especially in the case of the
"computed" mode), there is not a linear relationship between the value
of cnLineLabelDensityF and the number of labels. The curve of the
relationship is highly dependent on the particular data being plotted,
although I think it can safely be considered to be monotonic.

I can imagine users wanting to be able to specify programmatically that,
for example, there should be no less than 25 but no more than 40 line
labels in a plot. I have added another resource that supports an admittedly
clumsy method of doing this. After each draw, the get-only resource
cnLineLabelCount contains the number of line labels that were drawn. A
user could iteratively adjust  cnLineLabelDensityF based on the most
recent value of cnLineLabelCount, discarding plots until the specifications
were met.

For completeness, I also added similar resources cnHighLabelCount and
cnLowLabelCount for High and Low labels. There are not High and Low
label density resources however.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7049 10fe2890-8a40-4696-8b1d-7ccbb19c398e