equiv_sample_size did not originally return missing values. It does now, so C interface was modified to accomodate this.
The Fortran interface to eofcov_ts and eofcov_ts_pcmsg was modified to
add an additional flag that allows the user to specify whether the
mean is removed from the computation.  Thus, in the C interface, if a
"tsflag" attribute is set to 1, then the mean is not removed (the
default is to remove it).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7076 10fe2890-8a40-4696-8b1d-7ccbb19c398e