Fix for Sylvia's problem of trying to set explicit line labels in ContourPlot using an empty string.
Added a new resource Transformation class resource trGridType. This
specifies which Transformation sub-class is used to handle the
transformation of the data. It is mostly read-only, except when
a ScalarField or VectorField is set using 2D coordinate arrays. In
that case, it allows the user to choose between the Curvilinear and
the Spherical Transformations.

Also fixed some small descrepencies between the way the code was working
and the way it is documented.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7102 10fe2890-8a40-4696-8b1d-7ccbb19c398e