Added a new Workstation method UpdateDrawBB that allows a Workstation to update its cumulative bounding box with each object as it is drawn. It is implemented only for the PSWorkstation, which sends the bounding box to the LLU PS driver in ClearWorkstation method. A caveat is that no bounding box is calculated for graphics primitives -- this should not for Primitive objects which are drawn only in the data space of a Transform object, but it means that the bounding box may not be correct for immediate mode graphics primitives in all cases.
Also checking in code that allows basic access to PostScript hardware fonts.
This is very basic code. It performs no testing to prevent the user from
choosing fonts that are not available or characters that are not available
inside the fonts. The fonts are accessed by setting the FontQuality
resource to "Workstation". Also note that the characters are not
proportionally spaced.

Fixed the psResolution resource to truly have units of dots/inch as
advertised in the docs.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7185 10fe2890-8a40-4696-8b1d-7ccbb19c398e