This is to fix a problem with the routine CTTMRG that I've been working on for the past month.  The previous version of the routine could generate an incorrect triangular mesh for certain input rectangular grids (particularly those like the ORCA grid, for which the top row shrinks to a "polar line"). The problem didn't show up for the ORCA grid itself because all the points where the problem would have occurred were marked with a "missing data" value and that hid the problem from me.
I have a bit more testing to do on this, though I'm pretty confident that
what I've done does fix the problem and has no undesirable side effects,
because I've run what would seem to be the crucial tests.

I also fixed a problem in "yMakefile": there were references to a couple of
references to "CPxxxx" instead of "CTxxxx".


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7265 10fe2890-8a40-4696-8b1d-7ccbb19c398e