Consolidated some routines, like "set_xaxis_string" and "set_yaxis_string" into one routine, like "set_axis_string". Did same for is_valid_lat2d_attr/ is_valid_lon2d_attr, etc.
Also fixed it so that if data@description (or other valid "long_name"
type attributes) is set, then it is used for labeling the X and/or
Y axis.

PLEASE NOTE: as per request from Dennis and Sylvia, "data@units" is
also checked for, and is added to the X/Y axis string label.  This may
result in long and/or redundant titles, like:

     "Months since January (Months)"


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@7266 10fe2890-8a40-4696-8b1d-7ccbb19c398e